+++ src/tools.h
@@ -1,6 +1,7 @@
 #ifndef TOOLS_H
 #define TOOLS_H
 #include <stdlib.h>
-void u64_lshift1(u_int64_t *in, size_t n);
+#include <stdint.h>
+void u64_lshift1(uint64_t *in, uint64_t n);
 unsigned char set_bit(unsigned char in, unsigned short i, unsigned b);
 #endif // TOOLS_H
